==> 03_FIAS_clean_cities.ipynb <==
import pandas as pd
import numpy as np

pd.set_option('precision', 13)
pd.options.display.max_rows = 100

# загружаем города и поселки

work_cities = pd.read_csv('C:/00_Projects/Russia_color/all_cities_original.csv')

### Проверяем пустые значения ###

# смотрим пустые значения

work_cities[work_cities.offname.isnull()]

# в столбце city (бывшем formalname)

work_cities[work_cities.formalname.isnull()]

### Переименовываем столбцы ###

work_cities.columns

# оставляю нужные столбцы (беру formalname, в offname есть опечатки в названиях)

work_cities = work_cities[['aolevel', 'shortname', 'formalname',
                          'mun_district', 'region', 'okato', 'oktmo']]

# перименовываю столбцы (city у меня - любой населенный пункт)

work_cities = work_cities.rename(columns={
    'shortname': 'city_type',
    'formalname': 'city'})

## Удаляем дубликаты ##

Хотя мы удаляли дубликаты на предыдущем этапе, они все равно есть.

Но на всякий случай проверим, точно ли это идентичные записи.

**P.S.** У меня есть гипотеза, что это записи из разных таблиц. Т.е. некоторые записи дублируют в другой регион по ошибке или населенный пункт меняет принадлежность, но остается в таблице бывшего родительского региона с новым кодом.

### Проверяем ###

# заводим тестовый датафрейм

test = work_cities[work_cities.duplicated(keep=False)]

# смотрим индексы совпадающих строк

print(test.groupby(test.columns.tolist()).apply(lambda x: x.index.tolist()).values.tolist())

# нужно вставить индексы из прошлой ячейки

work_cities.loc[[160481, 162429]]

work_cities.loc[[2831, 5324]]

### Удаляем ###

work_cities.shape

work_cities = work_cities.drop_duplicates()

work_cities.shape

**Удаляем районы (муниципальные образования)**

Они остались с прошлого шага

# aolevel - 3 и city_type - р-н

work_cities = work_cities[~((work_cities.aolevel == 3) &
           (work_cities.city_type == 'р-н'))]

work_cities.shape

work_cities

**Меняем "г." на "г", "пгт." на "пгт" и "сп." на "сп"**

work_cities.city_type.replace('г.', 'г', inplace=True)
work_cities.city_type.replace('сп.', 'сп', inplace=True)
work_cities.city_type.replace('пгт.', 'пгт', inplace=True)

work_cities[work_cities.city_type == 'г.'].shape 

### Записываем в файл ###


work_cities.to_csv('C:/00_Projects/Russia_color/all_cities.csv',
                  sep=',', encoding='utf-8', index=False)

## Чистим города и поселки ##

work_cities.sample(10)

# смотрим сколько вообще  типов населенных пунктов есть в базе

work_cities.city_type.nunique()

# и каких

(work_cities
.groupby('city_type', as_index=False)
.agg({'aolevel': 'count'})
.rename(columns={'aolevel': 'city_number'})
.sort_values('city_type', ascending=False)
)

Удаляем то, что имеет адрес, но не является населенным пунктом: ж/д будки, автодороги, а также промзоны и территории.

# выгружаем все типы

print(work_cities.city_type.value_counts().index.to_list())

# отбираем ненужные

no_city = ['автодорога',
           'ж/д б-ка', 'ж/д пл-ка', 'ж/д_будка',
           'ж/д_казарм', 'ж/д_оп', 'ж/д_платф', 'ж/д_пост', 'ж/д_рзд',
           'казарма', 'п/о',
           'промзона', 'тер']

# Удаляем ненужные записи

work_cities = work_cities.loc[~work_cities.city_type.isin(no_city)]

work_cities.shape

### Записываем в файл ###

work_cities.to_csv('C:/00_Projects/Russia_color/all_cities_clean.csv',
                  sep=',', encoding='utf-8', index=False)
