==> 04_rus_colors_preprocessing.ipynb <==
import pandas as pd
import numpy as np

import matplotlib.pyplot as plt
import seaborn as sns

import re

%matplotlib inline

### Загружаем датасет ###

# загружаем все города и поселки

all_cities = pd.read_csv('C:/00_Projects/Russia_color/all_cities_clean.csv', encoding='utf-8')

all_cities

color_list = ['красн',
              'амарант', 'багрян', 'борд', 'вишн', 'гранат', 'рубин', # оттенки
              
              'розов',
              
              'желт', 'жёлт',              
              'бежев', 'лимон', 'янтар',
              
              'золот', 'алтан', 

              
              'зелен', 'зелён',
              'лайм', 'олива',
              
              'голуб', 'бирюз',
              
              'сини', 'синь', 'синяя', 'сине',
              'лазур',
              
              'фиолет', 'сирен',
              'аметист',
              
              'бел', 
              
              'черн', 'чёрн',
              
              'рыжо', 'рыже', 'рыжи', 'рыжа'
              'охра', 'охри',
              'ржав',
                            
              'серый', 'серо', 'сере', 'сери', 'серая', 
              'серебр', 'сиз'
             ]

unnecessary = [    
    'морозов', # розовый
    
    'коржав', # оранжевый, рыжый и ржавый
    'крыж',            
       
    'аксинин', # синий
    'аксинь',
    'есинеевка',
    'гусине',
    
    'кобел', # белый
    'колыбел',
    'аганино',
    'агафон',
    'агеев',
    'аграфен',
    'академ',
    

    # черный
    'караул',
    'печерник',
    'шарапов',
    
    # оттенки
    
    'зарубин', # рубиновый
]

color_list_non_slavic = [
    
    # тюркские, монгольские, финские
    
    'кызыл', 'кизил', 'улан', 'нарьян', 'чевер' # красный
    
    'цаган', 'саган', # белый
    
    'кара', 'хара', 'шеме', # черный
    
    'яшел', # зеленый
    
    # синий
    
    'сары', 'шара' # желтый
]

%%time

# циклом отбираем строки с нужными корнями

color_cities = pd.DataFrame()

for color in color_list:
    temp = all_cities[all_cities.city.str.contains(color, case=False)]
    color_cities = color_cities.append(temp, ignore_index = True)

color_cities.shape

# удаляю дубликаты (образуются из-за перечения корней)

color_cities.drop_duplicates(inplace=True)

color_cities.shape

%%time

# удаляю циклом строки с лишними корнями

for toponym in unnecessary:
    color_cities = color_cities[~(color_cities.city.str.contains(toponym, case=False))]

color_cities.shape

color_cities.sample(5)

#color_cities.to_excel('C:/Users/Boris/Desktop/work_cities.xlsx', index=False)

#color_cities.to_excel('C:/00_Projects/Russia_color/temp.xlsx', index=False)

%%time

# циклом отбираем строки с нужными корнями

color_cities_2 = pd.DataFrame()

for color in color_list_non_slavic:
    temp = all_cities[all_cities.city.str.contains(color, case=False)]
    color_cities_2 = color_cities_2.append(temp, ignore_index = True)
    
temp = all_cities[all_cities.city.str.contains('Ак', case=True)]
color_cities_2 = color_cities_2.append(temp, ignore_index = True)

temp = all_cities[all_cities.city.str.contains('Аг', case=True)]
color_cities_2 = color_cities_2.append(temp, ignore_index = True)

color_cities_2.shape

color_cities_2.drop_duplicates(inplace=True)

color_cities_2.shape

#color_cities_2.to_excel('C:/00_Projects/Russia_color/temp.xlsx', index=False)

### 100 самых распространенных названий ###

Если у поселков одного муниципального района одни и те же названия, в ФИАС указывается в скобках уточнение (обычно сельсовет).

Также в скобках иногда приводится второе название, но для нашей задачи они не важны (в топ-100 эти названия не входят).

# Количество названий с уточнениями

all_cities[all_cities.city.str.contains(' \(', case=False)].shape

# код удаляет уточнения (все, что после пробела и скобки)

(
    all_cities.city.replace(regex=True, to_replace=r' \(.*', value=r'')
    .replace(regex=True, to_replace=r'ё', value=r'е')
    .value_counts()
    .reset_index()
    .head(1000)
).to_excel('C:/00_Projects/Russia_color/temp.xlsx', index=False)
