==> 05_rus_colors.ipynb <==
import pandas as pd
import numpy as np

import matplotlib.pyplot as plt
import seaborn as sns

%matplotlib inline

### Загружаем датасет ###

# загружаю все города и поселки

all_cities = pd.read_csv('C:/00_Projects/Russia_color/all_cities_clean.csv', encoding='utf-8')

# загружаю все "цветные" города и поселки

color_cities = pd.read_excel('C:/00_Projects/Russia_color/work_cities_new.xlsx', sheet_name='need')

color_cities.shape

# ненужные цвета

unless_colors = [#'красный', 'золотой', 'зеленый',
               'вишневый', 'амарантовый',
               #'черный', 'белый',
               'багряный', 'бежевый',
               #'серый', 'ржавый',
               #'синий', 'голубой',
               #'желтый',
               'бордовый',
               #'серебряный',
               'гранатовый',
               #'лазурный',
               'лимонный', 'аметистовый',
               'оливковый',
               #'розовый',
               'охра', 'рубиновый',
               'сиреневый', 'рыжий',
               'сизый', 'янтарь', 'темный'
              ]

%%time

# оставляю только нужные цвета

for color in unless_colors:
    color_cities = color_cities[~(color_cities.color.str.contains(color, case=False))]

color_cities.shape

### Топ-5 цветов ###

# группирую славянские "прямые" топонимы

color_top = (color_cities
             [(color_cities.indirect == 0) &
              (color_cities.slavic == 1)]
             .groupby('color', as_index=False)
             .agg({'city':'count'})
             .rename(columns={'city':'stricts_slavic'})
             .sort_values('stricts_slavic', ascending=False)
            )

color_top.head()

# группирую неславянские "прямые" топонимы

non_slavic_strict = (color_cities
                     [(color_cities.indirect == 0) &
                      (color_cities.slavic == 0)]
                     .groupby('color', as_index=False)
                     .agg({'city':'count'})
                     .rename(columns={'city':'stricts_non_slavic'})
                     .sort_values('stricts_non_slavic', ascending=False)
                    )

non_slavic_strict.head()

# объединяю

color_top = color_top.merge(non_slavic_strict, how='left', on='color')

# группирую "косвенные" топонимы

all_indirects = (color_cities
                 [color_cities.indirect != 0]
                 .groupby('color', as_index=False)
                 .agg({'city':'count'})
                 .rename(columns={'city':'indirects'})
                 .sort_values('indirects', ascending=False)
                )

all_indirects.head()

# объединяю

color_top = color_top.merge(all_indirects, how='left', on='color')

# заполню NaN

color_top.fillna(0, inplace=True)

# суммирую по строкам

color_top['total'] = color_top.sum(axis=1)

# сортирую и записываю в файл

(color_top.sort_values('total', ascending=False)
# .to_excel('C:/00_Projects/Russia_color/temp.xlsx', index=False)
)

### Цвета по регионам ###

Для плитки

# считаю общее количество ойконимов по регионам

work = (all_cities
        .groupby('region', as_index=False)
        .agg({'city':'count'})
        .rename(columns={'city':'total_cities'})
        .sort_values('total_cities', ascending=False)
       )

work.head()

# считаю общее количество цветных городов (только прямые)

total_color_cities = (color_cities
                      [color_cities.indirect == 0]
                      .groupby('region', as_index=False)
                      .agg({'city':'count'})
                      .rename(columns={'city': 'total_color_cities'})
                     )

total_color_cities

# присоединяю к рабочему датафрейму

work = work.merge(total_color_cities, how='left', on='region')

work.shape

Считаю самый популярный цвет в регионе

# группирую по цветам и регионам

all_color_cities = (color_cities
                    [color_cities.indirect == 0]
                    .groupby(['region', 'color'], as_index=False)
                    .agg({'city':'count'})
                    .sort_values(['region', 'city'], ascending=False)
                    .rename(columns={'city': 'cities'})
                   )

(all_color_cities
 # .to_excel('C:/00_Projects/Russia_color/temp.xlsx', index=False)
).sample(5)

# оставляю только максимальные значения

# сначала группирую

max_color = (all_color_cities
             .groupby(['region'], as_index=False)
             .agg({'cities':'max'})
            )

# потом присоединяю

max_color = max_color.merge(all_color_cities, how='left', on=['region', 'cities'])

# присоединяю к рабочему фрейму

work = work.merge(max_color, how='left', on='region')

# удаляю регионы, где цветных городов меньше 10

work = work[~(work.total_color_cities < 10)]

work.region.value_counts()

# считаю вес

work['max_color_weight'] = work.cities / work.total_color_cities

#Загружаем названия регионов для datawrapper

region_name = pd.read_excel('C:/00_Projects/Russia_color/rus_color_viz.xlsx', sheet_name='reg_name')

work = work.merge(region_name, how='left', on='region')

all_color_cities

work

(work[['dw_region', 'region', 'total_color_cities', 'color', 'cities', 'max_color_weight']]
  .sort_values('max_color_weight', ascending=False)
 # .to_excel('C:/00_Projects/Russia_color/temp.xlsx', index=False)
).head()
