==> 06_mounths.ipynb <==
import pandas as pd
import numpy as np

import sys

import matplotlib.pyplot as plt
import seaborn as sns

%matplotlib inline

# загружаю все города и поселки

all_cities = pd.read_csv('C:/00_Projects/Russia_color/all_cities_clean.csv', encoding='utf-8')

list_mounth = [
    'январ', 'феврал',
    'март', 'апрел', 'май', 'мая', 'маев',
    'июн', 'июл', 'август',
    'сентяб', 'октябр', 'ноябр',
    'декабр'
]

%%time

# циклом отбираем строки с нужными корнями

mounths = pd.DataFrame()

for mounth in list_mounth:
    temp = all_cities[all_cities.city.str.contains(mounth, case=False)]
    temp['mounth'] = mounth
    mounths = mounths.append(temp, ignore_index = True)

mounths

#mounths.to_excel('C:/00_Projects/Russia_color/temp.xlsx', index=False)
